Split compiler in library and cli wrapper

+++ scripts/run.sh
@@ -1,3 +1,13 @@
 #!/usr/bin/env bash
 
-docker run --rm -it -v "$(pwd)"/types:/app/types wire-spec
+build_linux=false
+
+if [ "$WIRE_SPEC_BUILD_ALL" = true ] || [ "$WIRE_SPEC_BUILD_LINUX" = true ]; then
+  build_linux=true
+fi
+
+if [ $build_linux = true ]; then
+  docker run --rm -it -v "$(pwd)"/types:/app/types wire-spec
+  else
+    echo "WIRE_SPEC_BUILD_LINUX or WIRE_SPEC_BUILD_ALL not set to 'true'"
+fi
